[portafly] refactor imports
