+++ portafly/src/components/data-list/modals/ChangeStateModal.tsx
@@ -6,10 +6,9 @@ import {
   FormGroup
 } from '@patternfly/react-core'
 import { useTranslation } from 'i18n/useTranslation'
-import { DataListModal, useDataListBulkActions } from 'components/data-list'
+import { DataListModal, useDataListBulkActions, SubmitButton } from 'components/data-list'
 import { changeState } from 'dal/accounts/bulkActions'
-import { useAlertsContext } from 'components'
-import { SubmitButton } from './SubmitButton'
+import { useAlertsContext } from 'components/util'
 
 interface Props {
   items: string[]
